Extract shared status log dialog logic into useStatusLogDialog hook
Deduplicates ~80 lines of status dialog logic (draft state, validation,
submit handler, file upload, modal rendering) that was copied between
UpdateStatusButton and ColdRoomActionButton. Permission checking remains
in each caller for consistency.
